pcd-scan/requirements.txt
streamlit==1.28.0
opencv-python-headless==4.5.5.64
deepface==0.0.79
tensorflow-cpu==2.11.0
keras==2.11.0
Pillow
numpy
mediapipe==0.10.3
